+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### Bookmarks
 
-- Added ability to archive/unarchive bookmarks lists
+- **Added ability to archive/unarchive bookmarks lists**
 - Changed view title button to "Create Bookmarks List"
 - Collapse inactive lists by default, as well as non-virtual file bookmarks
+- Added icons for selection bookmarks
 
 ## 0.14
 
@@ -16,7 +17,7 @@
 
 ### Bookmarks
 
-- Introduced "Add File to Bookmarks" command and added an icon for it to the Editor title
+- **Introduced "Add File to Bookmarks" command and added an icon for it to the Editor title**
 - Allow bookmarking virtual files (non-bookmarked files that contain bookmarked selections)
 - Add "Change Current Bookmarks List" command
 - Selection bookmarks' values are now stored as is, without formatting or character limit